nvm on flask; use aiohttp, initial fetches work
I don't want to have to mess with all the threads that will be required
to make a flask streaming application work.  Async python seems like the
way to go here.

I liked simple_websockets, and tried to use quart or uvicorn so I could
use simple_websockets, but quart is broken in archlinux right now and
uvicorn is much too low-level for this prototype.  Aiohttp works fine
anyway, now that I worked out the details.

I've fixed the problems I identified ealier in the db design, namely
that threads were serving as both topics and threads.  I found another
threads issue, which is that threads should be identified by the comment
they branch off of, not as their own entity, otherwise mutliple people
responding to a single comment could create multiple threads that should
logically render as one single merged thread.  What a mess that would
be.

I also removed the many-to-many map between reports and topics for now,
since it isn't great for streaming (it's nice to have topics contain
that information).  AFICT it's only for opitimizing the reverse lookup
which is not something I want to bother with right now.  It's easy to
add back in later.

Streaming is in a "kinda sorta" working state.  A rough implementation
of the /ws protocol is ready, but there's no data to query yet and
there's also no NOTIFYs happening in the database yet.  So initial
fetches are always empty and live streaming is also always empty.

I'm going to stop and spend some time on fake data generator.