remove .user from FWUserAccount
Due to the complexity of trying to hide the extra layer of indirection
between the users table and the accounts table, the .user field in the
FWUserAccount table was just confusing what it stood for.

It should have stood for "what is the default user for this account".

But I myself got it wrong in not one, but two places.  And I don't see
why the UI should ever care about that concept.  So I removed it.